visual changes
I think I forgot a description last time, so I'll do it here.

I implemented the create_account function. It has the user go through a few inputs for name, email, password, etc. For each input, there are conditionals to make sure that the input is in the correct format.

Also implemented the helper method isInDataBase, which takes in an input and a field in UserInfo to search. It returns true if there is a match found